requirements.txt
# Chartastrophe - Essential dependencies
# Absurd correlation generator application

# Web framework
flask>=2.3.0
python-dotenv>=1.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
plotly>=5.15.0

# HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.0

# Statistical analysis
statsmodels>=0.14.0

# Structured logging
structlog>=24.1.0

# AI and NLP (for explanations)
transformers>=4.30.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0

# Security
cryptography>=42.0.0

# Development tools (uncomment if needed)
# black>=24.0.0
# flake8>=7.0.0
# pytest>=8.0.0
# pytest-cov>=4.1.0
# mypy>=1.8.0 